ready for build

+++ IrrationalNumbers/IrrationalNumbers.Logic/Expansions/ExponentTaylorExpansion.cs
@@ -12,9 +12,9 @@ namespace IrrationalNumbers.Logic.Expansions
 
             for (int i = 1; ; ++i)
             {
-                BigDecimal possibleRemainder = Math.Pow(3, c) / Utils.CalculateFactorial(i + 1) * Math.Pow(x, i + 1);
+                BigFloat possibleRemainder = Math.Pow(3, c) / Utils.CalculateFactorial(i + 1) * Math.Pow(x, i + 1);
 
-                if (BigDecimal.Abs(possibleRemainder) < Math.Pow(10, wantedRemainder))
+                if (BigFloat.Abs(possibleRemainder) < Math.Pow(10, wantedRemainder))
                     return new RemainderResult()
                     {
                         Remainder = possibleRemainder,
@@ -23,11 +23,11 @@ namespace IrrationalNumbers.Logic.Expansions
             }
         }
 
-        public BigDecimal ExpandFunction(int wantedRemainder, double x)
+        public BigFloat ExpandFunction(int wantedRemainder, double x)
         {
             RemainderResult remainderResult = EvaluateN(wantedRemainder, x);
 
-            BigDecimal result = Math.Pow(10,200);
+            BigFloat result = Math.Pow(10,200);
             for (int i = 1; i <= remainderResult.RemainderOrder; ++i)
             {
                 result = Math.Pow(x,i) / Utils.CalculateFactorial(i);
