+++ IrrationalNumbers/IrrationalNumbers.Logic/IBasicFunctionExpansion.cs
@@ -4,6 +4,6 @@ namespace IrrationalNumbers.Logic
 {
     public interface IBasicFunctionExpansion
     {
-        BigDecimal ExpandFunction(int wantedRemainder, double x);
+        BigFloat ExpandFunction(int wantedRemainder, double x);
     }
 }
