+++ IrrationalNumbers/IrrationalNumbers.Logic/RemainderResult.cs
@@ -6,6 +6,6 @@ namespace IrrationalNumbers.Logic
     public class RemainderResult
     {
         public int RemainderOrder { get; set; }
-        public BigDecimal Remainder { get; set; }
+        public BigFloat Remainder { get; set; }
     }
 }
