+++ IrrationalNumbers/IrrationalNumbers.Logic/Utils.cs
@@ -4,9 +4,9 @@ namespace IrrationalNumbers.Logic
 {
     public static class Utils
     {
-        public static BigDecimal CalculateFactorial(int n)
+        public static BigFloat CalculateFactorial(int n)
         {
-            BigDecimal result = 1;
+            BigFloat result = 1;
             for (int i = 1; i <= n; ++i)
                 result *= i;
             return result;
